README: more detail on the MGN9 rail length, and shorter rails
Discovered this the hard way.  I did not notice that the rail length was
longer than the v0 spec, so ordered an MGN9 rail the same length as my
previous MGN7 one.  This meant I was using the middle two screwholes
instead of three as described here in the BOM.  The deflection from the
the DragonBrick toolhead in this configuration was measured at 0.2° ±
0.1°.  Note these experiences in the relevant hint, and also make it
really clear that the rail length differs from the v0 BOM spec.